requirements.txt
# Core data science packages
numpy>=1.21.5
pandas>=1.3.5
scipy>=1.7.3

# Machine learning
torch>=1.11.0
#torchvision>=0.12.0
tensorflow>=2.9.1

# Data visualization
matplotlib>=3.5.2
#seaborn>=0.11.2

# Video processing and computer vision
opencv-python>=4.5.5.64
mediapipe>=0.8.9.1

# Audio preprocessing (Mel Spectrogram and denoising)
librosa>=0.10.2.post1
moviepy>=1.0.3

# Jupyter notebooks
jupyter>=1.0.0
ipython>=7.31.1

# Utility packages
tqdm>=4.64.0
pyyaml>=6.0

